+++ rcmt/source/github.py
@@ -1,7 +1,7 @@
 import datetime
 import fnmatch
 import io
-from typing import Any, Generator, TextIO, Union
+from typing import Any, Generator, Iterator, Optional, TextIO, Union
 
 import github
 import github.PullRequest
@@ -169,6 +169,15 @@ class Github(Base):
 
         SECRET_MASKER.add_secret(access_token)
 
+    def create_from_name(self, name: str) -> Optional[Repository]:
+        repo_name = "/".join(name.split("/")[1:])
+        try:
+            gh_repo = self.client.get_repo(full_name_or_id=repo_name, lazy=False)
+        except github.UnknownObjectException:
+            return None
+
+        return GithubRepository(access_token=self.access_token, repo=gh_repo)
+
     def list_repositories_with_open_pull_requests(
         self,
     ) -> Generator[Repository, None, None]:
@@ -178,16 +187,14 @@ class Github(Base):
         ):
             yield GithubRepository(self.access_token, issue.repository)
 
-    def list_repositories(self, since: datetime.datetime) -> list[Repository]:
+    def list_repositories(self, since: datetime.datetime) -> Iterator[Repository]:
         log.debug("start fetching repositories")
-        repos: list[Repository] = []
         for gh_repo in self.client.get_user().get_repos(
             direction="desc", sort="updated"
         ):
             if gh_repo.updated_at > since:
-                repos.append(GithubRepository(self.access_token, gh_repo))
+                yield GithubRepository(self.access_token, gh_repo)
             else:
                 break
 
         log.debug("finished fetching repositories")
-        return repos